+++ Runtime/BoolEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Runtime
+{
+    [CreateAssetMenu(menuName = "Channels/Bool Event Channel")]
+    public class BoolEventChannel : EventChannel<bool> { }
+}