+++ Runtime/DoubleEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Runtime
+{
+    [CreateAssetMenu(menuName = "Channels/Double Event Channel")]
+    public class DoubleEventChannel : EventChannel<double> { }
+}