+++ Runtime/EventChannel.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace Runtime
+{
+    public abstract class EventChannel : ScriptableObject
+    {
+        public event UnityAction EventRaised;
+
+        public void RaiseEvent() => EventRaised?.Invoke();
+    }
+
+    public abstract class EventChannel<T> : ScriptableObject
+    {
+        public event UnityAction<T> EventRaised;
+
+        public void RaiseEvent(T value) => EventRaised?.Invoke(value);
+    }
+}