+++ Runtime/FloatEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Runtime
+{
+    [CreateAssetMenu(menuName = "Channels/Float Event Channel")]
+    public class FloatEventChannel : EventChannel<float> { }
+}