+++ Runtime/IntEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Runtime
+{
+    [CreateAssetMenu(menuName = "Channels/Int Event Channel")]
+    public class IntEventChannel : EventChannel<int> { }
+}