+++ Runtime/TransformEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Runtime
+{
+    [CreateAssetMenu(menuName = "Channels/Transform Event Channel")]
+    public class TransformEventChannel : EventChannel<Transform> { }
+}