+++ Runtime/VoidEventChannel.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace Runtime
+{
+    [CreateAssetMenu(menuName = "Channels/Void Event Channel")]
+    public class VoidEventChannel : EventChannel { }
+}